Safari-611.1.10/LayoutTests/fast/text/web-font-load-invisible-during-loading-expected.txt

This test makes sure that fallback fonts are used during the time when fonts are loading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test1 is not americanTypewriterWidth
PASS test2 is americanTypewriterWidth
PASS test3 is americanTypewriterWidth
PASS test4 is americanTypewriterWidth
PASS successfullyParsed is true

TEST COMPLETE
This is rendered with American Typewriter.

This is rendered with American Typewriter.

This is rendered with American Typewriter.

This is rendered with American Typewriter.

This is rendered with American Typewriter.
